% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_predict.R
\name{item_predict}
\alias{item_predict}
\title{Generate Predictions for 1 User}
\usage{
item_predict(user, item_similarities, ratings_wide_read, predict_n)
}
\arguments{
\item{user}{An existing or new user denoted by their unique User.ID}

\item{item_similarities}{A wide matrix that contains cosine similarity scores
between each User and Item combination. A cosine similarity score of 1
indicates that Items are similar to each other and a cosine similarity
score of 0 indicates that Items are dissimilar (different from each other)
with numbers between 0 and 1 depicting relative similarity of Items. Items
are said to be similar if many Users have 'used' the Item.}

\item{ratings_wide_read}{A wide matrix which contains unique User ID's as row
names and unique Item ID's as column names with ratings as the values in
the matrix where 0 means not read and 1 means read.}

\item{predict_n}{The number of predictions to generate for the specified
user.}
}
\value{
The output dataset is a wide matrix that contains cosine similarity
  scores between each User and Item combination. A cosine similarity score of
  1 indicates that Items are similar to each other and a cosine similarity
  score of 0 indicates that Items are dissimilar (different from each other)
  with numbers between 0 and 1 depicting relative similarity of Items. Items
  are said to be similar if many Users have 'used' the Item.
}
\description{
Generate Predictions for 1 User
}
