% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_model_rmse.R
\name{mf_model_rmse}
\alias{mf_model_rmse}
\title{Matrix Factorisation Model RMSE for Cross Validation}
\usage{
mf_model_rmse(model, train_wide, test_wide)
}
\arguments{
\item{model}{The output to the `mf_model_build` function}

\item{train_wide}{The training dataset including rows with NAs respresenting the testing set.}

\item{test_wide}{The number of latent factors to add. This corresponds to the
number of columns in the W matrix and the number of rows in the H matrix}
}
\value{
The output is a dataframe of RMSE scores on the training, testing and
  complete set respectively.
}
\description{
Compare RMSE on the training and out of sample (testing) set
}
\details{
The input datasets are the wide training and the wide testing set.
  The output is a dataframe of RMSE scores on the training, testing and
  complete set respectively.
}
\examples{

}
