% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_model_selection.R
\name{mf_model_selection}
\alias{mf_model_selection}
\title{Matrix Factorisation Model Selection with Cross Validation}
\usage{
mf_model_selection(ratings_long, k_latent)
}
\arguments{
\item{ratings_long}{The input dataset is a long matrix which contains User ID's, Item ID's and Ratings in the first 3 columns respectively.}

\item{k_latent}{Set method = "user" to return cosine similarities between rows (users) eg. readers who gave similar books similar ratings would have cosine similarities closer to 1 and readers who give different ratings to similar books or who rated a different set of books have a cosine similarity closer to zero. Set method = "item" to return cosine similarities between columns (items) eg. users that gave movies similar ratings are considered similar and cosine similarity is closer to 1 and users that gave movies different ratings are considered different from each other and cosine similarity is closer to 0}
}
\value{
The output dataset is a wide matrix that contains cosine similarity scores between each unique User and Item combination. A cosine similarity score of 1 indicated that users or items are similar to each other and a cosine similarity score of 0 indicates that users or items are differnt from each other.
}
\description{
Train multiple models by varying parameters and compare RMSE on the training and out of sample (testing) set
}
\details{
The input dataset is a long matrix which contains User ID's, Item ID's and Ratings in the first 3 columns respectively.
The output is a grid of parameters that were varied and the RMSE on the training and test set respectively.
}
\examples{
cosine_similarities_user <- cosine_similarities(book_ratings, method = "user")
 cosine_similarities_item <- cosine_similarities(users_test, method = "item")

}
