% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_predict.R
\name{mf_predict}
\alias{mf_predict}
\title{Matrix Factorisation Predictions}
\usage{
mf_predict(model)
}
\arguments{
\item{model}{The model resulting from the matrix factorisation method which
is the output of the mf_model_build function in this package.}
}
\value{
The output is the predictions of the chosen model
}
\description{
Use the decomposed matrix to predict ratings
}
\details{
The predictions can be used to calculate the error and accuracy of
  the model's predictions on the training and or test (validation sets).
}
