% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_recommend.R
\name{mf_recommend}
\alias{mf_recommend}
\title{Generate Predictions for 1 User}
\usage{
mf_recommend(user, predictions, ratings_wide, predict_n)
}
\arguments{
\item{user}{An existing or new user denoted by their unique User.ID}

\item{predictions}{A wide matrix that contains predictions using the matrix
factorisation method. This is the output to the `mf_predict` function.}

\item{ratings_wide}{A wide matrix which contains unique User ID's as row
names and unique Item ID's as column names with ratings as the values in
the matrix. This is used to see whether a user has already 'used' an item
and exclude those from predictions.}

\item{predict_n}{The number of predictions to generate for the specified
user.}
}
\value{
The output dataset is a wide matrix that contains predictions of book
  ratings for users using the matrix factorisation method..
}
\description{
Generate Predictions for 1 User
}
