% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movie_ratings.R
\docType{data}
\name{movie_ratings}
\alias{movie_ratings}
\title{Movie Ratings Data}
\format{tibble}
\usage{
movie_ratings
}
\description{
20 movies seen or not seen by 15 movie-goers
}
\details{
This is a modified version of the "Movie Lens" dataset. The .RData file
  contains a small subset of 15 movie-goers (Users) and 20 movies (Items) and whether they have
  seen (rating=1) or not seen (rating=0) each movie. The full dataset has
  ratings on 40 000 movies by 260 000 users, some 24 million ratings in all.
  There are four csv files: `links.csv`, `movies.csv`, `ratings.csv`, and
  `tags.csv`.
}
