% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_blank_na.R
\name{remove_blank_na}
\alias{remove_blank_na}
\title{Remove Missing Values or NAs}
\usage{
remove_blank_na(ratings_long)
}
\arguments{
\item{ratings_long}{The input dataset contains 3 columns for User ID, Item ID
and Rating respectively}
}
\value{
A long dataset without zero ratings and without NA ratings
}
\description{
Takes a long dataset and removes implicit ratings (not rated ie.
rating value = zero) and NA values if applicable
}
\details{
Implicit ratings are those that occur in a dataset with a value of
zero eg. the movie has been seen but not rated or the book has been read
but not rated. In user based and item based collaborative filtering it is
necessary to remove the zero ratings because missing values are replaced
with zeros in order for the cosine similarity calculation to work. In
contrast, the matrix factorisation technique replaces missing values with
NAs so that the model doesn't learn to predict a zero but learns a more
appropriate and useful value
}
