% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Training and Testing (Validation) Split}
\usage{
train_test_split(ratings_long, split)
}
\arguments{
\item{ratings_long}{The input dataset contains 3 columns for User ID, Item ID
and Rating respectively}

\item{split}{This is the percentage of the dataset assigned to the training
set eg. 0.8 denotes 80%% assigned to training and 20%% assigned to testing}
}
\value{
The output is 2 datasets, train and test, in wide format
}
\description{
Randomly splits the dataset into training and test sets based on the proportion specified
}
\details{
The input dataset contains 3 columns for User ID's, Item ID's and
Ratings respectively. The implicit (zero) ratings should be removed before
splitting into training and testing but this is catered for in this
function in case this is forgotten. The output is 2 datasets, train and
test, in wide format
}
