% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movie_ratings.R
\docType{data}
\name{viewed_movies}
\alias{viewed_movies}
\title{Movie Ratings Data}
\format{tibble}
\usage{
viewed_movies
}
\description{
Ratings (on a scale of 0 to 10) of 150 books from 10,000 users.
}
\details{
This is a modified version of the "Movie Lens" dataset.
(\link{https://grouplens.org/datasets/movielens/)}. The .RData file
contains a small subset of 15 users and 20 movies and whether they have
seen (rating=1) or not seen (rating=0) each movie. The full dataset has
ratings on 40 000 movies by 260 000 users, some 24 million ratings in all.
There are four csv files: \code{links.csv}, \code{movies.csv}, \code{ratings.csv}, and
\code{tags.csv}.
}
