% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_model_build.R
\name{mf_model_build}
\alias{mf_model_build}
\title{Matrix Factorisation Model Build}
\usage{
mf_model_build(train_wide, k_latent, alpha, beta, bias)
}
\arguments{
\item{train_wide}{The input dataset is a long matrix which contains User ID's, Item ID's and Ratings in the first 3 columns respectively.}

\item{k_latent}{Set method = "user" to return cosine similarities between rows (users) eg. readers who gave similar books similar ratings would have cosine similarities closer to 1 and readers who give different ratings to similar books or who rated a different set of books have a cosine similarity closer to zero. Set method = "item" to return cosine similarities between columns (items) eg. users that gave movies similar ratings are considered similar and cosine similarity is closer to 1 and users that gave movies different ratings are considered different from each other and cosine similarity is closer to 0}
}
\value{
The output is the predictive model which can then be used to make predictions and calculate the accuracy as measured by the RMSE..
}
\description{
Build the matrix factorisation model
}
\details{
The input dataset is a long matrix which contains User ID's, Item ID's and Ratings in the first 3 columns respectively.
The output is the model specified which can then be used to predict ratings..
}
